refactor: remove (routes) folder from user api endpoints
